refactor: extract Monaco SQL editor into reusable component
